DummyBackend: Use 64-bit interface

+++ device_backends/include/DummyBackend.h
@@ -57,8 +57,8 @@ namespace ChimeraTK {
      */
     void close() override;
 
-    void read(uint8_t bar, uint32_t address, int32_t* data, size_t sizeInBytes) override;
-    void write(uint8_t bar, uint32_t address, int32_t const* data, size_t sizeInBytes) override;
+    void read(uint64_t bar, uint64_t address, int32_t* data, size_t sizeInBytes) override;
+    void write(uint64_t bar, uint64_t address, int32_t const* data, size_t sizeInBytes) override;
 
     bool isFunctional() const override { return (_opened && !_hasActiveException); }
